refactor: AppLayout
1. Layout 하위로 이동
#25
