+++ components/common/Layout/types/index.ts
@@ -1,0 +1,3 @@
+export interface Children {
+  children: React.ReactNode;
+}
